+++ backend/app/serializers/role.py
@@ -3,7 +3,7 @@ from rest_framework import serializers
 from garpix_company.models.user_role import get_company_role_model
 
 
-class GarpixCompanyRoleSerializer(serializers.ModelSerializer):
+class CompanyRoleSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = get_company_role_model()