+++ backend/garpix_company/permissions/company_admin.py
@@ -1,8 +1,10 @@
 from rest_framework import permissions
 
 from garpix_company.models import get_company_model, UserCompany, InviteToCompany
+from garpix_company.models.user_role import get_company_role_model
 
 Company = get_company_model()
+CompanyRole = get_company_role_model()
 
 
 class CompanyAdminOnly(permissions.BasePermission):
@@ -15,9 +17,9 @@ class CompanyAdminOnly(permissions.BasePermission):
             return True
 
         if isinstance(obj, Company):
-            return request.user.is_authenticated and request.user.id in obj.usercompany_set.filter(
-                is_admin=True).values_list('user', flat=True)
+            return request.user.is_authenticated and request.user.id in obj.user_companies.filter(
+                role=CompanyRole.get_admin_role()).values_list('user', flat=True)
         if isinstance(obj, InviteToCompany) or isinstance(obj, UserCompany):
-            return request.user.is_authenticated and obj.company.usercompany_set.filter(
-                is_admin=True, user=request.user)
+            return request.user.is_authenticated and obj.company.user_companies.filter(
+                role=CompanyRole.get_admin_role(), user=request.user)
         return False