+++ backend/garpix_company/serializers/invite.py
@@ -9,11 +9,18 @@ from garpix_company.models.invite import InviteToCompany
 from django.utils.translation import ugettext_lazy as _
 
 
+# RoleSerializer = import_string(getattr(settings, 'GARPIX_COMPANY_ROLE_SERIALIZER', 'garpix_company.serializers.GarpixCompanyRoleSerializer'))
+from garpix_company.models.user_role import get_company_role_model
+
+
 class InviteToCompanySerializer(serializers.ModelSerializer):
 
     class Meta:
         model = InviteToCompany
-        fields = ('email', 'is_admin', 'role')
+        fields = ('email', 'role')
+        extra_kwargs = {
+            'role': {'required': True},
+        }
 
     def validate_email(self, value):
         User = get_user_model()
@@ -21,25 +28,25 @@ class InviteToCompanySerializer(serializers.ModelSerializer):
         try:
             user = User.objects.get(email=value)
             if not Company.check_user_companies_limit(user):
-                raise ValidationError(_(f'У пользователя с указанным email превышен лимит количества компаний'))
+                raise ValidationError(_('У пользователя с указанным email превышен лимит количества компаний'))
         except User.DoesNotExist:
             raise ValidationError(_('Пользователь с указанным email не зарегистрирован'))
         return value
 
     def create(self, validated_data):
-        Company = get_company_model()
-        # getting user
-        user = None
+        Role = get_company_role_model()
         request = self.context.get("request")
         company_id = self.context.get("company_id")
+        if role := validated_data.get('role', None):
+            if role.role_type == Role.ROLE_TYPE.OWNER:
+                raise ValidationError({'role': [_('Нельзя пригласить пользователя на роль владельца')]})
         if request and hasattr(request, "user") and request.user.is_authenticated:
             with transaction.atomic():
                 # creating
                 obj = InviteToCompany(
                     email=validated_data['email'],
                     company_id=company_id,
-                    is_admin=validated_data['is_admin'],
-                    role=validated_data.get('role', None)
+                    role=role
                 )
                 obj.save()
                 return obj
@@ -50,25 +57,31 @@ class CreateAndInviteToCompanySerializer(serializers.ModelSerializer):
 
     class Meta:
         model = InviteToCompany
-        fields = ('email', 'is_admin', 'role')
+        fields = ('email', 'role')
+        extra_kwargs = {
+            'role': {'required': True},
+        }
 
     def validate_email(self, value):
         User = get_user_model()
         if User.objects.filter(email=value).first():
-            ValidationError(_('Пользователь с указанным email уже зарегистрирован'))
+            raise ValidationError(_('Пользователь с указанным email уже зарегистрирован'))
         return value
 
     def create(self, validated_data):
         User = get_user_model()
+        Role = get_company_role_model()
         request = self.context.get("request")
         company_id = self.context.get("company_id")
+        if role := validated_data.pop('role', None):
+            if role.role_type == Role.ROLE_TYPE.OWNER:
+                raise ValidationError({'role': [_('Нельзя пригласить пользователя на роль владельца')]})
         if request and hasattr(request, "user") and request.user.is_authenticated:
             with transaction.atomic():
                 invite_data = {
                     'email': validated_data['email'],
                     'company_id': company_id,
-                    'is_admin': validated_data.pop('is_admin'),
-                    'role': validated_data.pop('role', None)
+                    'role': role
                 }
                 if 'username' not in validated_data.keys():
                     validated_data['username'] = get_random_string(25)
@@ -79,3 +92,12 @@ class CreateAndInviteToCompanySerializer(serializers.ModelSerializer):
                 obj.save()
                 return obj
         return None
+
+
+class InvitesSerializer(serializers.ModelSerializer):
+
+    # role = RoleSerializer()
+
+    class Meta:
+        model = InviteToCompany
+        fields = '__all__'